requirements.txt
pandas==2.2.2
openpyxl==3.1.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34
python-dotenv==1.0.1
duckdb==1.0.0
